refactor: api로 navershop 등록
